refactor(db): Renames DBManager to MongoDbManager
