+++ server.js
@@ -15,7 +15,7 @@ const { Tracker } = require('node-moving-things-tracker');
 const cloneDeep = require('lodash.clonedeep');
 const Opendatacam = require('./server/Opendatacam');
 const { getURLData } = require('./server/utils/urlHelper');
-const DBManager = require('./server/db/DBManager');
+const DBManager = require('./server/db/MongoDbManager');
 const FileSystemManager = require('./server/fs/FileSystemManager');
 const { MjpegProxy } = require('./server/utils/mjpegproxy');
 const config = require('./config.json');
